Fix int to float conversion warnings across project
Add explicit static_cast<float>() conversions to eliminate all implicit
int-to-float conversion warnings found during compilation.

Fixed locations:
1. IconSpawnerSystem.cpp:86-87 - Grid position calculations
   - float x = columnGrid * iconWidth (int * int → float)
   - float y = rowGrid * iconHeight (int * int → float)

2. IconSpawnerSystem.cpp:52 - Timer increment
   - timer += TIME_PER_FRAME.asMilliseconds() (float += int)

3. BGEntity.cpp:24 - Initial position setup
   - setPosition(0, -WINDOW_HEIGHT * 7) (int, int → float, float)

4. BGEntity.cpp:39 - Position reset in update loop
   - setPosition(0, -WINDOW_HEIGHT * 7) (int, int → float, float)

All conversions are now explicit and safe, eliminating compiler warnings
while maintaining correct behavior.

+++ src/entities/BGEntity.cpp
@@ -21,7 +21,7 @@ void BGEntity::initialize() {
     this->sprite->setTexture(*texture);
 
     this->sprite->setTextureRect(IntRect(0, 0, BaseRunner::WINDOW_WIDTH, BaseRunner::WINDOW_HEIGHT * 8));
-    this->setPosition(0, -BaseRunner::WINDOW_HEIGHT * 7);
+    this->setPosition(0.0f, static_cast<float>(-BaseRunner::WINDOW_HEIGHT * 7));
 }
 
 void BGEntity::processInput(Event event) {
@@ -36,5 +36,5 @@ void BGEntity::update(Time deltaTime) {
     auto localPos = this->sprite->getPosition();
 
     if (localPos.y * deltaTime.asSeconds() > 0)
-        this->setPosition(0, -BaseRunner::WINDOW_HEIGHT * 7);
+        this->setPosition(0.0f, static_cast<float>(-BaseRunner::WINDOW_HEIGHT * 7));
 }
